requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2020.12.5
    # via requests
chardet==3.0.4
    # via requests
idna==2.10
    # via requests
lxml==4.6.1
    # via nitter (setup.py)
pyyaml==5.3.1
    # via nitter (setup.py)
requests==2.25.0
    # via nitter (setup.py)
sdnotify==0.3.2
    # via nitter (setup.py)
urllib3==1.26.3
    # via requests
